Add tests to ai fix assessments upsell class and remove old feature upsell
